APP: update filter shorted url

+++ src/app/user/user.component.ts
@@ -7,22 +7,23 @@ import {GetUsersResponse} from "../share/model/user/get-users-response.model";
 import {ResponseInfo} from "../share/model/common/response-info.model";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSort} from "@angular/material/sort";
-import {catchError, map, merge, startWith, switchMap, throwError} from "rxjs";
+import {map, merge, startWith, switchMap} from "rxjs";
 import {GetUsersRequest} from "../share/model/user/get-users-request.model";
 import {RequestInfo, SortInfo} from "../share/model/common/request-info.model";
 import {MatTableDataSource} from "@angular/material/table";
-import {CheckboxFilterComponent} from "../share/UI/custom-filter/checkbox-filter.component";
+import {CheckboxFilterComponent} from "../share/UI/custom-filter/check-box-filter/checkbox-filter.component";
 import {HeaderTitleService} from "../header/header-title.service";
-import {QueryParams} from "../share/model/common/query-params.model";
 import {dataSourceActiveFilter, dataSourceRoleFilter} from "../share/constant/data-source-filter.constant";
 import {Role} from "../share/constant/role.constant";
-import {RadioBoxFilterComponent} from "../share/UI/custom-filter/radio-box-filter.component";
+import {RadioBoxFilterComponent} from "../share/UI/custom-filter/radio-box-filter/radio-box-filter.component";
 import {IdResponse} from "../share/model/common/id-response.model";
 import {Alert} from "../share/constant/alert.constant";
 import {MatDialog} from "@angular/material/dialog";
 import {PopConfirmComponent} from "../share/UI/pop-comfirm/pop-confirm/pop-confirm.component";
 import {PopConfirmModel} from "../share/model/UI/pop-confirm.model";
 import {PopConfirmConstant} from "../share/constant/pop-confirm.constant";
+import {MatDateRangePicker} from "@angular/material/datepicker";
+import {QueryParamsUserConstant as paramUser} from "../share/constant/query-params-user.constant";
 
 @Component({
   selector: 'app-user',
@@ -30,9 +31,10 @@ import {PopConfirmConstant} from "../share/constant/pop-confirm.constant";
   styleUrls: ['./user.component.scss'],
 })
 export class UserComponent implements OnInit, AfterViewInit {
-  protected readonly ADMIN = Role.ADMIN;
-  protected readonly MODERATOR = Role.MODERATOR;
-  protected readonly MEMBER = Role.MEMBER;
+  protected readonly AD = Role.AD;
+  protected readonly MOD = Role.MOD;
+  protected readonly MEM = Role.MEM;
+  protected readonly paramUser = paramUser;
   protected readonly dataSourceRoleFilter = dataSourceRoleFilter;
   protected readonly dataSourceActiveFilter = dataSourceActiveFilter;
 
@@ -40,7 +42,7 @@ export class UserComponent implements OnInit, AfterViewInit {
   responseInfo: ResponseInfo;
   isLoading: boolean = false;
   dataSource: MatTableDataSource<User>;
-  queryParams: QueryParams;
+  queryParams: any;
   isRecycleMode: boolean = false;
 
   @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
@@ -64,7 +66,7 @@ export class UserComponent implements OnInit, AfterViewInit {
     this.dataSource = new MatTableDataSource<User>(this.users);
     this.dataSource.paginator = this.paginator;
     this.dataSource.sort = this.sort;
-    this._route.queryParams.subscribe((params: QueryParams) => {
+    this._route.queryParams.subscribe((params) => {
       this.queryParams = params;
     });
   }
@@ -75,21 +77,15 @@ export class UserComponent implements OnInit, AfterViewInit {
       this.paginator.page,
       this.roleFilter.dropdownFilter.dropdownHidden,
       this.activeFilter.dropdownFilter.dropdownHidden,
+      // this.picker.closedStream,
     ).pipe(
       startWith({}),
       switchMap(() => {
         this.isLoading = true;
         this._cd.detectChanges();
         let getUsersRequest = this.buildGetUserRequest();
-        return this._userService.getUsers(getUsersRequest)
-          .pipe(catchError((err) => throwError(() => {
-              this._alertService.handleErrors(err)
-            }
-          )));
-      }),
-      map(data => {
-        return data;
-      }),
+        return this._userService.getUsers(getUsersRequest);
+      })
     ).subscribe({
       next: (resData: GetUsersResponse) => {
         this.users = resData.data;
@@ -97,6 +93,7 @@ export class UserComponent implements OnInit, AfterViewInit {
         this.isLoading = false;
         this._cd.detectChanges();
       },
+      error: err => this._alertService.handleErrors(err)
     });
   }
 
@@ -104,10 +101,12 @@ export class UserComponent implements OnInit, AfterViewInit {
     let sortInfo = new SortInfo(this.sort.direction, this.sort.active);
     let requestInfo = new RequestInfo(this.paginator.pageIndex, this.paginator.pageSize, sortInfo);
     let getUsersRequest = new GetUsersRequest(requestInfo);
-    getUsersRequest.setRole = this.queryParams.role;
-    getUsersRequest.isActive = this.queryParams.active;
+    getUsersRequest.setRole = this.queryParams[paramUser.ROLE];
+    getUsersRequest.isActive = this.queryParams[paramUser.ACTIVE];
     getUsersRequest.isExcludeCurrentUserLogged = true;
     getUsersRequest.isDeleted = this.isRecycleMode;
+    getUsersRequest.fromDate = new Date(this.queryParams[paramUser.FROM_DATE]);
+    getUsersRequest.toDate = new Date(this.queryParams[paramUser.TO_DATE]);
     return getUsersRequest;
   }
 
@@ -140,8 +139,8 @@ export class UserComponent implements OnInit, AfterViewInit {
       data: new PopConfirmModel("Are you sure to delete this user"),
     });
     dialogRef.afterClosed().subscribe(action => {
-      switch (action) {
-        case PopConfirmConstant.TEXT_SURE: {
+      if (action === PopConfirmConstant.TEXT_OK) {
+        {
           this.isLoading = true;
           this._userService.deleteUser(id).subscribe({
             next: (resData: IdResponse) => {
@@ -179,7 +178,7 @@ export class UserComponent implements OnInit, AfterViewInit {
   onToggleRecycleMode() {
     this.isRecycleMode = !this.isRecycleMode;
     this.loadUsers();
-    let titlePage = this.isRecycleMode? "USER RECYCLE BIN" : "USER MANAGEMENT";
+    let titlePage = this.isRecycleMode ? "USER RECYCLE BIN" : "USER MANAGEMENT";
     this._headerService.setTitle(titlePage);
   }
 
@@ -212,7 +211,7 @@ export class UserComponent implements OnInit, AfterViewInit {
     this.isLoading = true;
     this._userService.restoreUser(id).subscribe({
       next: (resData: IdResponse) => {
-        if(resData.id) {
+        if (resData.id) {
           this.loadUsers();
         }
         this._alertService.success(Alert.USER_RESTORE_SUCCESS);
